feat: show spinner until loading tos
