+++ kernel/static/index.html
@@ -502,8 +502,8 @@
     <div class="eth-login-popup">
       <img class="eth-login-logo" src="images/decentraland-connect/Isologotipo.png"></img>
       <div class="eth-login-description">Enter the first virtual world fully owned by its users.</div>
-      <div id="eth-login-confirmation-spinner" class="loader" style="display: none;"></div>
-      <div id="eth-login-confirmation-wrapper">
+      <div id="eth-login-confirmation-spinner" class="loader" style="display: block;"></div>
+      <div id="eth-login-confirmation-wrapper" style="display: none;">
         <div class="eth-login-tos">
           <input type="checkbox" id="agree-check" class="eth-login-tos-agree" onchange="checkTos()"></input>
           <label for="agree-check" class="eth-login-tos-label">I am of legal age and I have read and agree to the <a href="https://decentraland.org/terms" target="_blank">Terms of Service</a> and <a href="https://decentraland.org/privacy" target="_blank">Privacy Policy</a></label>
